Add files via upload
In the pov.jpg we can see that when the 100th eigenvalues is included, the pov is greater than 0.9. In that case, that we should preserve 100 dimensions in order to explain at least 90% of the variance.
 
The best number of clusters is:  5
All 4 socres are the highest among all clusters 
As we loop the centers for 1000 times for 000 data, we will get close to the true centers
The ARI and v-measure score are close to 0.9, which our prediction is very close to the true clusters.
The MRI score also highest at 1.4
the silhouette coefficient was highest when k = 5, suggesting 5 is the optimal number of clusters. 

There are 10000 samples and 10 number of features

The scores for all clusters are similar, there is no specific best number of clusters.
These may caused by the fact that in the image set, the number are similar and not clearly separated.